+++ huskarl-core/src/client_auth/client_secret.rs
@@ -73,7 +73,7 @@ impl ClientAuthentication for ClientSecret {
         &'a self,
         client_id: &'a str,
         _issuer: Option<&'a str>,
-        _token_endpoint: &'a Uri,
+        _endpoint: &'a Uri,
         allowed_methods: Option<&'a [String]>,
     ) -> MaybeSendBoxFuture<'a, Result<AuthenticationParams<'a>, Error>> {
         Box::pin(async move {